fix bug

+++ Final_Project/Final_Project/ConsoleGUI.cpp
@@ -112,7 +112,7 @@ void DrawTextFile(const char* filename, int x, int y)
 	file.close();
 }
 
-void printtext(string str, int x, int y) {
+void printtext(string str, double x, double y) {
 	gotoxy(x, y);
 	cout << str << endl;
 }
